Create a .gitignore for the repository and add initial edits to .vimrc
Adds a .gitignore file to the repository's top-level directory, which (in short) will ignore all files and
directories in the top-level dirs, 'colors/' and 'bundle/'. HOWEVER, this file DOES NOT IGNORE THE DIRECTORIES
THEMSELVES! I thought it would be unnecessary to include all of the bundles and colorschemes I use with my vim
configuration since most are free, open-source and available on github, and thus can easily be restored if
need be. I left the empty directories, however, in case I decide to override the .gitignore for select files,
which could prove difficult if the directory itself was also being ignored by git.

On branch master
Changes to be committed:
    modified:   .config/.vimrc
    modified:   .gitignore