Fix errors in .vimrc and complete .vimrc up to 'Key Mappings'
The values for part of the 'statusline' option and the 'wildmode' option were entered incorrectly
during my previous session working on .config/.vimrc and thus, were not applied to the vim configuration.
These errors were corrected and the option settings are now applied properly when using vim. Minimal changes
were made other than the corrections (one augroup, which included only two autocmd statements) in order to
complete a "working" 1st draft of .config/.vimrc for all sections up to (but not including) "Key Mappings".

Changes to be committed:
    modified:   .vimrc