Add initial basic changes necessary for migration to neovim
This commit is fairly lightweight. A couple minor changes to .config/.vimrc for better compatibility between both vim
and neovim as I work on transitioning as well as the new 'spell/' directory, which is mostly ignored (see matching
changes in '.gitignore') other than the *.spl and *.sug files for the 'en' locale and the *.diff file for the 'en_US'
locale.

More changes in next commit as I start on some ftplugins and whatnot for my ruby development environment.

Changes to be committed:
	modified:   .config/.vimrc
	modified:   .gitignore
	new file:   spell/README_en.txt
	new file:   spell/en.ascii.spl
	new file:   spell/en.ascii.sug
	new file:   spell/en.latin1.spl
	new file:   spell/en.latin1.sug
	new file:   spell/en.utf-8.spl
	new file:   spell/en.utf-8.sug
	new file:   spell/en_US.diff