Add completed 'autoload/rubycomplete.vim' file
This commit just updates the, now completed, rubycomplete.vim file, which is located in the 'autoload/' directory. As
a side note, this file is merely a fork (with only extremely minor semantic & commentary changes) of a file of the same
name found in the vim-ruby repository on github. The URL of the repository as well as the URL of the original
'rubycomplete.vim' file are listed below and I take no credit, either explicit or implied, for this script or any others
that I have or will use from the repository, all of which will be noted as such.

Original Repo:      vim-ruby/vim-ruby
Repo Address:       https://github.com/vim-ruby/vim-ruby

vim-ruby/.../rubycomplete.vim URL:
                    https://github.com/vim-ruby/vim-ruby/blob/master/autoload/rubycomplete.vim

Changes to be committed:
    modified:   autoload/rubycomplete.vim