Add completed 'indent/ruby.vim' file
Adds the completed file 'indent/ruby.vim', which defines the function for determining the level of indentation for
a given line when the 'filetype' option has the value, 'ruby'. Credits for this script go to vim-ruby/vim-ruby:
        https://github.com/vim-ruby/vim-ruby/blob/master/indent/ruby.vim

Hopefully I can have some of the syntax/ or spec/ directories completed by the next commit.