Initial Commit for dev/4ruby remote branch. Add syntax/ruby.vim
Just a housekeeping note, which is mostly for my own memory: this is the first commit for the local branch,
dev/vim-ruby, as well as for the associated remote-tracking branch, origin/dev/4ruby, with origin being my remote github
repo located at:
                    https://github.com/StevenWard94/myvim

I plan to use this branch for the remainder of my "work" on setting up a (somewhat...) functional/helpful ruby on rails
environment to work in.
------------------------------------------------------------------------------------------------------------------------

As for the actual change in content of this commit, it will commit the completed version of the syntax/ruby.vim file,
credits to which go to https://github.com/vim-ruby/vim-ruby.

Next, I may forego (at least for now) making a 'syntax/eruby.vim' file, since I really don't plan on working too much
with eruby, so I would rather complete a few more necessary files (like those for the currently empty 'spec/' directory)
first to expedite my having a working rails development environment.