Add haskell-complete 'config/.vimrc.bundles' file
The current version of 'config/.vimrc.bundles' is complete and functional but it only configures plugins for use with
GHC and Haskell scripts. Will update for ruby, C/C++ and html/css in the near future (hopefully...). This commit also
adds the 'snippets/' directory, contained in which is the 'haskell.snippets' file; obviously a file for vim-snippets
completion of haskell code blocks.