Add completed version of 'syntax/haskell.vim
The newly complete file, 'syntax/haskell.vim' is a syntax highlighting file for haskell scripts that adds some useful
behaviors to vim's default haskell syntax file (the exact additions are itemized at the top of the file). Aside from
a few small changes -- one of which being the addition of folding -- credit for this file's contents goes to Stephen
Diehl, although I do intend on making a few more alterations to the file as my personal preferences are more clearly
determined. The original file can be found at the URL below:

                https://github.com/sdiehl/haskell-vim-proto/blob/master/vim/syntax/haskell.vim