Add more plugins and Fix :nohlsearch mapping
Added a number of plugins to config/.vimrc.bundles and fixed the mapping of <leader>/ so
that it, correctly, executes :nohlsearch instead of :invhlsearch. This new mapping lets
vim start highlighting searches again the next time I start a search, instead of making
me use the mapping again to turn the search highlighting back on.