Add 'plugin/' directory with 'home_library.vim' file
The new 'plugin/' directory will contain my personal scripts, downloaded plugins, and
any other '*.vim' files I want loaded but that are not really a "bundle" and probably
wouldn't be loaded properly by vim-plug in the 'bundle/' directory. Currently, the
directory contains only the 'home_library.vim' script, which is very short as of right
now. I plan to add quite a bit more to it in the near future and may not use all of its
commands that frequently, so it uses the format shown in the *write-plugin-quickload*
section of the vim help-docs to load commands on startup and then if/when a command is
used, a 'FuncUndefined' event triggers an autocmd to load the remainder of the script
(i.e. the actual definitions of the functions called by the previously-defined user
commands). This will prevent noticeable slowing of vim's startup time if this (or any)
plugin script gets to be very large.