Add syntax file to conceal haskell syntax with nice operators
This commit adds the 'after/syntax/haskell.vim' file, which lists a number of syntax
matches for the 'haskell' filetype. These matches will be concealed by digraph operators
commonly used in mathematics that have the meaning of the (often times infix) function
or haskell operator. This is purely cosmetic and just makes haskell scripts a bit easier
to read.