Auto headers for C & unfold json
Created the necessary functions to add a mapping when editing C and C++
files that automatically inserts a multi-line header comment with the
author's name, the file name, and the date of the last change made
(defaults to current date). If the function detects that the file is
within a git repository, it will also include the url in the comment.
This makes it a lot easier to quickly create documentation for the top
of the file. This commit also changed the default 'conceallevel' for
json files to 0 so that quotation marks are not hidden.

Example header comment for a new 'Node.hpp' file contained in the 'include'
directory of the 'data-structures' git repository (Note: the number
of '*' characters is determined by the current value for 'textwidth'):
    /***************************************************************//**
     * Author:       Steven Ward <stevenward94@gmail.com>
     * File:         data-structures/include/Node.hpp
     * URL:          https://github.com/StevenWard94/data-structures
     * Last Change:  2022 Oct 23
     ****************************************************************/

Example header comment for a new 'calculate.cpp' file saved in my home
directory:
    /***************************************************************//**
     * Author:       Steven Ward <stevenward94@gmail.com>
     * File:         /home/steven/calculate.cpp
     * Last Change:  2022 Oct 23
     ****************************************************************/